move mui component to NavBar component
